cleaning up testing of varnish context hash
